Massively simplify codebase by leveraging Koog framework
