--- core/CLAUDE.md
@@ -1,1 +1,0 @@
-README.md
