--- core/agents/api/build.gradle.kts
@@ -1,21 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-}
-
-group = "com.duchastel.simon.brainiac.core.modelproviders"
-
-kotlin {
-
-    jvm()
-
-    linuxX64()
-    macosX64()
-    macosArm64()
-    mingwX64()
-
-    sourceSets {
-        val commonMain by getting {
-
-        }
-    }
-}
