--- core/fileaccess/api/build.gradle.kts
@@ -1,24 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.kotlin.serialization)
-}
-
-group = "com.duchastel.simon.brainiac.core.fileaccess"
-
-kotlin {
-    jvm()
-
-    linuxX64()
-    macosX64()
-    macosArm64()
-    mingwX64()
-
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                implementation(libs.kotlinx.serialization.core)
-                implementation(libs.okio)
-            }
-        }
-    }
-}
