--- core/fileaccess/api/src/commonMain/kotlin/AccessLogEntry.kt
@@ -1,16 +1,0 @@
-@file:OptIn(ExperimentalTime::class)
-
-package com.duchastel.simon.brainiac.core.fileaccess
-
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.Contextual
-import kotlin.time.Instant
-import kotlin.time.ExperimentalTime
-
-@Serializable
-data class AccessLogEntry(
-    @Contextual val timestamp: Instant,
-    val action: String,
-    val path: String,
-    val details: String? = null
-)
