--- core/fileaccess/impl/build.gradle.kts
@@ -1,49 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.kotlin.serialization)
-    alias(libs.plugins.ksp)
-    alias(libs.plugins.kotest)
-}
-
-group = "com.duchastel.simon.brainiac.core.fileaccess"
-
-kotlin {
-    jvm()
-
-    linuxX64()
-    macosX64()
-    macosArm64()
-    mingwX64()
-
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(project(":core:fileaccess:api"))
-                implementation(libs.kotlinx.serialization.core)
-                
-                implementation(libs.kotlinx.coroutines.core)
-                implementation(libs.okio)
-                implementation(libs.kaml)
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                implementation(libs.kotest.framework.engine)
-                implementation(libs.kotest.assertions.core)
-                
-            }
-        }
-        val jvmTest by getting {
-            dependencies {
-                implementation(libs.kotest.runner.junit5)
-            }
-        }
-    }
-}
-
-tasks.named<Test>("jvmTest") {
-    useJUnitPlatform()
-    filter {
-        isFailOnNoMatchingTests = false
-    }
-}
