--- core/identity/README.md
@@ -1,7 +1,0 @@
-# Core Identity Module
-
-Centralized service for managing the AI system's core identity.
-
-## Service Architecture
-
-- **CoreIdentityService**: Interface for accessing the core identity
