--- core/identity/api/src/commonMain/kotlin/CoreIdentityService.kt
@@ -1,7 +1,0 @@
-package com.duchastel.simon.brainiac.core.identity
-
-interface CoreIdentityService {
-    fun getCoreIdentity(): CoreIdentity
-
-    fun getCoreIdentityContent(): String
-}
