--- core/identity/impl/src/commonMain/kotlin/DefaultCoreIdentityService.kt
@@ -1,19 +1,0 @@
-package com.duchastel.simon.brainiac.core.identity
-
-import com.duchastel.simon.brainiac.core.fileaccess.FileSystemService
-import okio.Path
-
-class DefaultCoreIdentityService(
-    private val fileSystemService: FileSystemService,
-    private val coreIdentityPath: Path
-) : CoreIdentityService {
-
-    override fun getCoreIdentity(): CoreIdentity {
-        val content = fileSystemService.read(coreIdentityPath)
-        return CoreIdentity(content)
-    }
-
-    override fun getCoreIdentityContent(): String {
-        return fileSystemService.read(coreIdentityPath)
-    }
-}
