--- core/process/core-loop/api/src/commonMain/kotlin/CoreLoopProcess.kt
@@ -1,5 +1,0 @@
-package com.duchastel.simon.brainiac.core.process
-
-interface CoreLoopProcess {
-    suspend fun processUserPrompt(userPrompt: String): String
-}
