--- core/process/search/AGENTS.md
@@ -1,1 +1,0 @@
-README.md
