--- core/process/search/README.md
@@ -1,19 +1,0 @@
-# Core Search Module
-
-Memory retrieval and indexing functionality for efficient long-term memory access in the Brainiac AI Memory System.
-
-## Why Dedicated Search Module
-
-- **Semantic Retrieval**: Enables context-aware memory search beyond simple keyword matching
-- **Hierarchical Indexing**: Leverages directory structure and `_index.md` files for organized memory access
-- **Performance**: Optimized search algorithms for large memory repositories
-- **Relevance Ranking**: Intelligent ordering of search results based on context and access patterns
-
-## Key Design Decisions
-
-- **Index-Based Search**: Uses `_index.md` manifests and relationship links for semantic navigation
-- **Multi-Modal Queries**: Supports both content-based and metadata-based search criteria
-- **Access Pattern Integration**: Search results inform and are informed by memory access logging
-- **Context Sensitivity**: Search queries generated from working memory context for optimal relevance
-
-This module implements the memory retrieval mechanisms that enable the Core Loop to efficiently locate relevant long-term memories for context assembly, supporting the specification's requirement for intelligent memory access.
