--- core/process/search/api/src/commonMain/kotlin/SearchService.kt
@@ -1,7 +1,0 @@
-package com.duchastel.simon.brainiac.core.search
-
-import com.duchastel.simon.brainiac.core.fileaccess.LTMFile
-
-interface SearchService {
-    suspend fun searchLTM(query: String): List<LTMFile>
-}
