+++ settings.gradle.kts
@@ -1,14 +1,3 @@
 rootProject.name = "brainiac"
 
-include(
-    "agents",
-    "core:identity:api",
-    "core:identity:impl",
-    "core:fileaccess:api",
-    "core:fileaccess:impl",
-    "core:agents:api",
-    "core:process:core-loop:api",
-    "core:process:core-loop:impl",
-    "core:process:search:api",
-    "core:process:search:impl",
-)
+include("core-loop")
